Delete index.html
AOL Mail offers its customers free mail services and it is convenient to use. You can easily create a new account in the blink of an eye. It is highly user-friendly and gives us a personalized experience. A division of Yahoo, AOL is providing a free web-based email service known as the “AOL Mail”, where AOL stands for American Online.